LDAP UpdateGroupsService: again, do not fill the log with useless notifications.
This time a log message with "groups [] were remove" is triggered any
time when no groups are to be removed.

+++ apps/user_ldap/lib/Service/UpdateGroupsService.php
@@ -199,6 +199,9 @@ class UpdateGroupsService {
 	 */
 	public function handleRemovedGroups(array $removedGroups): void {
 		$this->logger->debug('service "updateGroups" - dealing with removed groups.');
+		if (empty($removedGroups)) {
+			return;
+		}
 
 		$this->groupMembershipMapper->deleteGroups($removedGroups);
 		foreach ($removedGroups as $group) {
